# proxy/public/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of Geoservice API.",
    "title": "Geoservice.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/api/address/geocode": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "address"
        ],
        "summary": "Вычисление адресса по широте и долготе.",
        "operationId": "GeocodeRequest",
        "parameters": [
          {
            "description": "Lat - широта\nLng - долгота",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GeocodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GeocodeResponse"
          }
        }
      }
    },
    "/api/address/search": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "address"
        ],
        "summary": "Вычисление местанахождения по адрессу.",
        "operationId": "SearchRequest",
        "parameters": [
          {
            "description": "Qury - запрос, представляющий собой адрес",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SearchRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SearchResponse"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Авторизация пользователя.",
        "operationId": "LoginRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Регистрация пользователя.",
        "operationId": "RegisterRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/RegisterReponse"
          }
        }
      }
    },
    "/api/user/list": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Получить список пользователей.",
        "operationId": "ListRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/ListResponse"
          }
        }
      }
    },
    "/api/user/profile": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Получить профиль пользователя по email.",
        "operationId": "ProfileRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProfileRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ProfileResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "Address": {
      "type": "object",
      "properties": {
        "lat": {
          "type": "string",
          "x-go-name": "Lat"
        },
        "lon": {
          "type": "string",
          "x-go-name": "Lon"
        }
      },
      "x-go-package": "proxy/internal/models"
    },
    "GeocodeRequest": {
      "type": "object",
      "properties": {
        "lat": {
          "type": "string",
          "x-go-name": "Lat"
        },
        "lng": {
          "type": "string",
          "x-go-name": "Lng"
        }
      },
      "x-go-package": "proxy/internal/modules/geo/controller"
    },
    "GeocodeResponse": {
      "type": "object",
      "properties": {
        "addresses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          },
          "x-go-name": "Addresses"
        }
      },
      "x-go-package": "proxy/internal/modules/geo/controller"
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "proxy/internal/modules/auth/controller"
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "proxy/internal/modules/auth/controller"
    },
    "ProfileRequest": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string"
        }
      },
      "x-go-package": "proxy/internal/modules/user/controller"
    },
    "RegisterReponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "proxy/internal/modules/auth/controller"
    },
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "phone": {
          "type": "string",
          "x-go-name": "Phone"
        }
      },
      "x-go-package": "proxy/internal/modules/auth/controller"
    },
    "SearchRequest": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string",
          "x-go-name": "Query"
        }
      },
      "x-go-package": "proxy/internal/modules/geo/controller"
    },
    "SearchResponse": {
      "type": "object",
      "properties": {
        "addresses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          },
          "x-go-name": "Addresses"
        }
      },
      "x-go-package": "proxy/internal/modules/geo/controller"
    },
    "User": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        }
      },
      "x-go-package": "proxy/internal/models"
    }
  },
  "responses": {
    "GeocodeResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GeocodeResponse"
      }
    },
    "ListResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "LoginResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/LoginResponse"
      }
    },
    "ProfileResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/User"
      }
    },
    "RegisterReponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/RegisterReponse"
      }
    },
    "SearchResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/SearchResponse"
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}